pricing-engine/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
numpy==2.1.3
scipy==1.14.1
lightgbm==4.5.0
httpx==0.28.0
pydantic==2.10.3
python-dotenv==1.0.1
